changed the positions of some functions
