+++ tests/test_interface/test_magi_attn_interface.py
@@ -25,7 +25,6 @@ from torch.testing._internal.common_distributed import skip_if_lt_x_gpu
 from torch.testing._internal.common_utils import run_tests
 
 import magi_attention
-import magi_attention.testing
 from magi_attention.api.functools import (
     apply_padding,
     compute_pad_size,
